|#3|@FESG3002| controle ajouter sur la fonction clear directory pour eviter tout risque

+++ tests/test_unitaire_speed_up.py
@@ -42,7 +42,7 @@ def ensure_dir_exists(directory):
 
 def clean_directory(directory):
     """Nettoie un dossier et tous ses contenus."""
-    if os.path.exists(directory):
+    if os.path.exists(directory) and (" " in directory)==False: #verifier qui ya aucune espace dans le destination du fichier
         shutil.rmtree(directory)
 
 @pytest.fixture(autouse=True)
